flask/requirements.txt
google-generativeai
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.17.1
opencv-python==4.8.1.78
numpy==1.26.4
werkzeug==2.3.7
gunicorn==21.2.0
dotenv
